refactor: move divider system into its own module (src/divider.rs)
First step of decomposing the god trio: the Divider enum, full_island_w /
DIFF_GUTTER_W geometry, and the 5 divider methods move out of main.rs +
app.rs into src/divider.rs (a crate-root child module, so its `impl Kyde`
block still reaches Kyde's private fields like render.rs does). Re-exported
from the crate root so existing call sites are untouched. ~155 lines off the
god files, zero behaviour change.
